+++ operator/pkg/resources/target/statefulset.go
@@ -100,7 +100,7 @@ func targetPodSpec(ais *aisv1.AIStore) *corev1.PodSpec {
 		},
 		Containers: []corev1.Container{
 			{
-				Name:            "ais-node",
+				Name:            cmn.AISContainerName,
 				Image:           ais.Spec.NodeImage,
 				ImagePullPolicy: corev1.PullAlways,
 				Command:         []string{"aisnode"},